+++ src/main/java/lotto/LottoResultUseCase.java
@@ -1,0 +1,5 @@
+package lotto;
+
+public interface LottoResultUseCase {
+
+}
