--- src/main/java/lotto/view/converter/MessageConverter.java
@@ -1,8 +1,0 @@
-package lotto.view.converter;
-
-import java.util.List;
-
-public interface MessageConverter<T> {
-
-    List<String> convert(List<T> targets);
-}
