+++ src/main/java/lotto/view/converter/MessageParser.java
@@ -1,0 +1,10 @@
+package lotto.view.converter;
+
+import java.util.List;
+
+public interface MessageParser<T> {
+
+    List<String> toMessages(List<T> targets);
+
+    List<Integer> toNumbers(String message);
+}
